Create food control module
this should results in a food on the map, which has 8s for visibility and 3s for invisibility. it should appear in random positions and had a seed for unique LFSP, which would allow different foods in different appearing pattern, to avoid overlap in the same position. (Although still possible, but really rare to happen now.)